backend/requirements.txt
Django>=2.1,<2.2
psycopg2
channels
channels_redis
djangorestframework
watchdog
django-silk
coverage
